First naive approach for dollar only values
Split the number into chunks of hundreds and then calculate the hundreds-number repesentation. Then add the current scale (millions, thousands) and concatenate everything.
The word representation hundreds-number itself has some tricky pitfalls, like the numbers between 10 and 19 and, that zero is only a mentioned number, when there is nothing else